fix: 13ver permission대응 및 14ver permission 수정
