+++ app/src/main/AndroidManifest.xml
@@ -4,6 +4,7 @@
 
 
     <uses-feature android:name="android.hardware.camera.any" />
+
     <uses-permission android:name="android.permission.RECORD_AUDIO" />
 
     <uses-permission android:name="android.permission.INTERNET" />
@@ -13,14 +14,20 @@
     <uses-permission android:name="android.permission.INTERNET" />
 
     <!--    갤러리 권한-->
-    <uses-permission android:name="android.permission.READ_EXTERNAL_STORAGE"
+    <uses-permission
+        android:name="android.permission.READ_EXTERNAL_STORAGE"
         android:maxSdkVersion="32" />
-    <uses-permission android:name="android.permission.WRITE_EXTERNAL_STORAGE"
+    <uses-permission
+        android:name="android.permission.WRITE_EXTERNAL_STORAGE"
         android:maxSdkVersion="28" /> <!-- Android 10 이하에서만 유효 -->
 
     <!-- Android 12 및 이후 버전용 퍼미션 설정 -->
-    <uses-permission android:name="android.permission.READ_MEDIA_IMAGES"
+    <uses-permission
+        android:name="android.permission.READ_MEDIA_IMAGES"
         android:minSdkVersion="31" />
+
+    <!--    Android 14 이후 버전용 일부 엑세스 권한-->
+    <uses-permission android:name="android.permission.READ_MEDIA_VISUAL_USER_SELECTED" />
 
     <!--    카메라 권한-->
     <uses-permission android:name="android.permission.CAMERA" />
